Clear data from the passthrough cache for complex or long running operations

Clear data from the passthrough cache for complex or long running operations

Usage

cache_clear(
  .cache = rappdirs::user_cache_dir(utils::packageName()),
  .prefix = ".*",
  interactive = TRUE
)

Arguments

.cache

the location of the cache as a directory. May get its value from options("ggrrr.cache.dir") or the default value of rappdirs::user_cache_dir("ggrrr")

.prefix

a regular expression matching the prefix of the cached item, so that do selective clean up operations. defaults to everything.

interactive

suppress `are you sure?` warning with a FALSE value (defaults to TRUE)

Value

nothing. called for side effects