Download a shape file map

This function downloads and caches a zipped shape file map, unpacks it, finds the correct shape file layer optionally simplifies it and converts it to `sf`.

Usage

downloadMap(zipUrl, mapName = NULL, simplify = FALSE, ...)

Arguments

zipUrl

the URL of the zipped shape file

mapName

(optional) the layer name or map name - this is `"xyz"` if a zip file contains `"xyz.shp"`. Usually there is one `.shp` file in a zip file, and by default it will be picked if this is not set.

simplify

do you want to simplify the map?

...

Named arguments passed on to .cached

.nocache

an option to defeat the caching which can be set globally as options("cache.disable"=TRUE)

.stale

the length of time in days to keep cached data before considering it as stale. can also be set by options("cache.stale")

Value

a `sf` object containing the map

Examples

if (FALSE) { # \dontrun{
downloadMap(
  zipUrl="https://bit.ly/3A9TnR1"
)
} # }