Get field information from a feature service layer.

Get field information from a feature service layer.

Usage

featureServerLayerInfo(url, ...)

Arguments

url

the feature service layer url. If no layer is specified the first one is selected.

...

Named arguments passed on to .cache_download

.nocache

if set to TRUE all caching is disabled

.cache

the location of the downloaded files

.stale

how long to leave this file before replacing it.

.extn

the file name extension

Value

a dataframe of field names for the selected layer

Examples

svc = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/ArcGIS/rest/services"
fs = sprintf("%s/%s",svc,"Countries_December_2024_Boundaries_UK_BUC/FeatureServer")
layer = sprintf("%s/%s",fs,"0")
featureServerLayerInfo(layer)
#> # A tibble: 11 × 9
#>    name         type  alias sqlType nullable editable domain defaultValue length
#>    <chr>        <chr> <chr> <chr>   <lgl>    <lgl>    <lgl>  <chr>         <int>
#>  1 FID          esri… FID   sqlTyp… FALSE    FALSE    NA     NA               NA
#>  2 CTRY24CD     esri… CTRY… sqlTyp… TRUE     TRUE     NA     NA                9
#>  3 CTRY24NM     esri… CTRY… sqlTyp… TRUE     TRUE     NA     NA               16
#>  4 CTRY24NMW    esri… CTRY… sqlTyp… TRUE     TRUE     NA     NA               17
#>  5 BNG_E        esri… BNG_E sqlTyp… TRUE     TRUE     NA     NA               NA
#>  6 BNG_N        esri… BNG_N sqlTyp… TRUE     TRUE     NA     NA               NA
#>  7 LONG         esri… LONG  sqlTyp… TRUE     TRUE     NA     NA               NA
#>  8 LAT          esri… LAT   sqlTyp… TRUE     TRUE     NA     NA               NA
#>  9 Shape__Area  esri… Shap… sqlTyp… TRUE     FALSE    NA     NA               NA
#> 10 Shape__Leng… esri… Shap… sqlTyp… TRUE     FALSE    NA     NA               NA
#> 11 GlobalID     esri… Glob… sqlTyp… FALSE    FALSE    NA     NEWID() WIT…     38