Create a new map with a popout panel to show areas where there is a high density of people for example.

Defaults work well for London on an England only map.

Usage

popoutArea(
  shape,
  popoutShape = arear::londonShape,
  popoutPosition = c("NE", "NW", "SE", "SW"),
  popoutScale = 3,
  nudgeX = 0.25,
  nudgeY = 0.25
)

Arguments

shape

The original shape

popoutShape

The mask shape. This will be summarised before applied.

popoutPosition

Which corner to place the popout NE,NW,SE or SW

popoutScale

A factor to grow the popout area by. This is linear scale so the popout will appear the square of this factor bigger.

nudgeX

shift the popout panel by a small amount (in coordinate units)

nudgeY

shift the popout panel by a small amount (in coordinate units)

Value

A new map with the content intersecting the popout area duplicated, expanded and placed in the specified corner.