Locations of UK general medical hospitals in mid march 2020 with estimates of beds available and maximal surge capacity HDU beds

This was manually assembled and curated from various sources in mid march 2020 as the NHS geared up to provide additional capacity to cope with the surge in COVID cases. It is not an up to date picture of NHS capacity. It does not include mental health or community hospitals. The surge capacity seems to have been calculated quite differently in Scotland.

Usage

surgecapacity

Format

A sf geometry with:

nation

England, Wales, etc...

hospitalId

An id for the hospital

sector

NHS or independent

hospitalName

the hospital name

pcds

the UK postcode of the hospital

trustId

the NHS trust or local health board of the hospital

trustName

the NHS trust or local health board name

tier1

indicator of the role of the hospital as an acure provider

hduBeds

the number of hdu beds the hospital could have provided at maximum surge in March 2020

acuteBeds

the number of acute beds the hospital could have provided at maximum surge in March 2020