Mid year 2019 small area estimates for adult population of England, Wales, Scotland, and Northern Ireland

Small area single digit estimates are aggregated to include only over 65s, and combining gender.

Usage

uk2019retiredpopulation

Format

A 41730 line data frame:

code

the ONS code for the ares

name

The name for the area

codeType

the type of area (LSOA, DZ or LGD)

population

the size of the population

Source

England and Wales

https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates

Scotland

https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/population/population-estimates/2011-based-special-area-population-estimates/small-area-population-estimates/time-series#2018

Northern Ireland

https://www.opendatani.gov.uk/dataset/3333626e-b96e-4b90-82fb-474c6c03b868/resource/64bd8dc4-935f-4bdd-9232-90ff33f24732/

Details

Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/