COVID reporting map

Shape file related to the detailed level for PHE coronavirus statistical reporting in England, Wales, Scotland, and Northern Ireland on the coronavirus.gov.uk tracker. The legacy is the level at which detailed data was reported in the first wave, prior to the API and was aggregated by https://github.com/tomwhite/covid-19-uk-data

Usage

ukcovidmap(legacy = FALSE)

Source

England, Wales, Scotland & NI - LAD19

https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc

England & Wales - LAD19

https://geoportal.statistics.gov.uk/datasets/local-authority-districts-december-2019-boundaries-uk-buc

Scotland - CA19

https://www.spatialdata.gov.scot/geonetwork/srv/eng/catalog.search#/metadata/1cd57ea6-8d6e-412b-a9dd-d1c89a80ad62

Northern Ireland -LGD12

https://data.gov.uk/dataset/05f72866-b72b-476a-b6f3-57bd4a768674/osni-open-data-largescale-boundaries-local-government-districts-2012

Arguments

legacy

use the legacy version of the map from pre coronavirus api times.

Value

A 382 line data frame (or a 335 line data frame for legacy):

code

the ONS code for the ares

name

The name for the area

codeType

the type of area (LAD19)

altCode

NA

Details

This map matches the data reported by the PHE coronavirus api when it is downloading lower tier local authority regions e.g. via https://api.coronavirus.data.gov.uk/v2/data?areaType=ltla&metric=newCasesBySpecimenDate&format=csv The API has now been discontinued.

There are 2 regions here for which no data is reported in this API - the city of London & the Isles of Scilly.

Data available under open government licence: http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/